# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/portal_audit_logs:
    get:
      tags:
        - PortalAuditLogs
      summary: List PortalAuditLogs
      description: Get a paginated list of PortalAuditLogs with optional filtering and sorting
      operationId: listPortalAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalAuditLogList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PortalAuditLogs
      summary: Create PortalAuditLog
      description: Create a new PortalAuditLog
      operationId: createPortalAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalAuditLogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalAuditLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/portal_invites:
    get:
      tags:
        - PortalInvites
      summary: List PortalInvites
      description: Get a paginated list of PortalInvites with optional filtering and sorting
      operationId: listPortalInvites
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalInviteList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PortalInvites
      summary: Create PortalInvite
      description: Create a new PortalInvite
      operationId: createPortalInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalInviteRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalInvite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    PortalAuditLog:
      type: object
      required:
        - id
        - event
        - occurred_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        event:
          type: string
        portal_user_id:
          nullable: true
          type: string
          format: uuid
        invite_id:
          nullable: true
          type: string
          format: uuid
        token_id:
          nullable: true
          type: string
          format: uuid
        actor:
          nullable: true
          type: string
          maxLength: 120
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreatePortalAuditLogRequest:
      type: object
      required:
        - event
        - occurred_at
        - metadata
      properties:
        event:
          type: string
        portal_user_id:
          nullable: true
          type: string
          format: uuid
        invite_id:
          nullable: true
          type: string
          format: uuid
        token_id:
          nullable: true
          type: string
          format: uuid
        actor:
          nullable: true
          type: string
          maxLength: 120
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    PortalAuditLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PortalAuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PortalInvite:
      type: object
      required:
        - id
        - recipient_email
        - token_nonce
        - token_expires_at
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        recipient_email:
          type: string
          maxLength: 320
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        status:
          type: string
        granted_by:
          nullable: true
          type: string
          format: uuid
        redeemed_by:
          nullable: true
          type: string
          format: uuid
        redeemed_at:
          nullable: true
          type: string
          format: date-time
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreatePortalInviteRequest:
      type: object
      required:
        - recipient_email
        - token_nonce
        - token_expires_at
        - status
        - metadata
      properties:
        recipient_email:
          type: string
          maxLength: 320
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        status:
          type: string
        granted_by:
          nullable: true
          type: string
          format: uuid
        redeemed_by:
          nullable: true
          type: string
          format: uuid
        redeemed_at:
          nullable: true
          type: string
          format: date-time
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    PortalInviteList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PortalInvite'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []